+++ tools/test-js-with-node
@@ -124,7 +124,6 @@ EXEMPT_FILES = make_set(
         "web/src/message_view_header.js",
         "web/src/message_viewport.js",
         "web/src/messages_overlay_ui.ts",
-        "web/src/muted_topics_ui.js",
         "web/src/muted_users_ui.js",
         "web/src/narrow.js",
         "web/src/navbar_alerts.js",
@@ -230,6 +229,7 @@ EXEMPT_FILES = make_set(
         "web/src/user_status.js",
         "web/src/user_status_ui.js",
         "web/src/user_topics.js",
+        "web/src/user_topics_ui.js",
         "web/src/webpack_public_path.js",
         "web/src/zcommand.js",
         "web/src/zform.js",
